Split structs used in oc image append into separate files
